Testing placing files
C:\Users\Cesar\Documents\GitHub\team-b\Documentation
Above appears to be my local repository path, so I placed a file, and it
looks like it was uploaded and visible to the team.
But, I just place it in my local pc. I haven't uploaded into Github, so
I don't know how this is visible now.