Added files via upload
Team, I have uploaded a sample code of our login. I tested it out and it works.  I have also attached a .txt file in the "documentation" folder, which I used to test this code and it worked.  It basically allows the user to enter their name and password for a total of 3 times before the program terminates.